move waveforms to their own file, add PiecewiseLinear waveform, add evalDeriv to the waveforms (for convolution with impulse response) remove extraneous options from em1d
